Ml_model/requirements.txt
flask
pdfplumber
pymupdf
scikit-learn
pandas
numpy
joblib
transformers
torch
pytesseract
Pillow
